Makefile: Bump version, separate DMGs for .pkg and .app
You can install either from the DMG with the .pkg file
or with the DMG with the .app bundle.

The PKG file doesn't seem to work as well with macOS 14, so I only
recommend it on older systems, and, even then, it's not my first
recommendation.  I might delete the .pkg installer targets eventually...